Move EraBasedProtocolParametersUpdate to Cardano.Api.Compatible.ProtocolParametersUpdate
Relocates the era-based protocol parameter update GADT, its era-sliced
record types, createEraBasedProtocolParamUpdate, createPParams and the
pre-Conway UpdateProposal machinery into a new Compatible module. The
UpdateProposal cluster has to move with the GADT: UpdateProposal contains
it, and its FromCBOR instance depends on fromLedgerPParamsUpdate, which in
turn uses the private pparamsUpdateTo* helpers.

Cardano.Api.ProtocolParameters re-exports the new module, so its export
list and Cardano.Api are unchanged. Cardano.Api.Compatible re-exports the
new module as well.
