+++ cardano-api/src/Cardano/Api/Compatible.hs
@@ -1,6 +1,8 @@
 module Cardano.Api.Compatible
-  ( module Cardano.Api.Compatible.Tx
+  ( module Cardano.Api.Compatible.ProtocolParametersUpdate
+  , module Cardano.Api.Compatible.Tx
   )
 where
 
+import Cardano.Api.Compatible.ProtocolParametersUpdate
 import Cardano.Api.Compatible.Tx
